+++ src/apps/halofi/halofi.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 
 import { AbstractApp } from '~app/app.dynamic-module';
 
+import { BaseHalofiGameContractPositionFetcher } from './base/halofi.game.contract-position-fetcher';
 import { CeloHalofiGameContractPositionFetcher } from './celo/halofi.game.contract-position-fetcher';
 import { HalofiGameGamesApiSource } from './common/halofi.game.games.api-source';
 import { HalofiContractFactory } from './contracts';
@@ -16,6 +17,8 @@ import { PolygonHalofiGameContractPositionFetcher } from './polygon/halofi.game.
     PolygonHalofiGameContractPositionFetcher,
     // Celo
     CeloHalofiGameContractPositionFetcher,
+    // Base
+    BaseHalofiGameContractPositionFetcher,
   ],
 })
 export class HalofiAppModule extends AbstractApp() {}