+++ src/apps/y2k-finance/arbitrum/y2k-finance.mint-v1.contract-position-fetcher.ts
@@ -58,21 +58,21 @@ export class ArbitrumY2KFinanceMintV1ContractPositionFetcher extends ContractPos
   async getTokenDefinitions(
     params: GetTokenDefinitionsParams<Y2KFinanceVaultV1, DefaultContractPositionDefinition>,
   ): Promise<UnderlyingTokenDefinition[] | null> {
-    const epochIds = await this.getEpochIds(params.multicall, params.contract);
+    const epochIdsRaw = await this.getEpochIds(params.multicall, params.contract);
     const claimableAsset = await params.contract.asset();
+    const epochIds = epochIdsRaw.map(x => x.toString());
     return epochIds
-      .map(id => [
+      .map(tokenId => [
         {
           metaType: MetaType.SUPPLIED,
           address: params.contract.address,
           network: this.network,
-          tokenId: id.toNumber(),
+          tokenId,
         },
         {
           metaType: MetaType.CLAIMABLE,
           address: claimableAsset,
           network: this.network,
-          tokenId: id.toNumber(),
         },
       ])
       .flat();
@@ -92,7 +92,9 @@ export class ArbitrumY2KFinanceMintV1ContractPositionFetcher extends ContractPos
     const vault = params.multicall.wrap(params.contract);
     const results = await Promise.all(
       epochIds.map(async id => {
+        const finalTVL = await vault.idFinalTVL(id);
         const balance = await vault.balanceOf(params.address, id);
+        if (finalTVL.isZero() || balance.isZero()) return [0, 0];
         const claimable = await vault.previewWithdraw(id, balance);
         return [balance, claimable];
       }),